+++ crates/model/src/events/order/snapshot.rs
@@ -67,6 +67,8 @@ pub struct OrderSnapshot {
     pub quantity: Quantity,
     /// The order price (LIMIT).
     pub price: Option<Price>,
+    /// The order activation price for trailing-stop orders.
+    pub activation_price: Option<Price>,
     /// The order trigger price (STOP).
     pub trigger_price: Option<Price>,
     /// The trigger type for the order.
@@ -148,6 +150,7 @@ impl From<OrderAny> for OrderSnapshot {
             quantity: order.quantity(),
             price: order.price(),
             trigger_price: order.trigger_price(),
+            activation_price: order.activation_price(),
             trigger_type: order.trigger_type(),
             limit_offset: order.limit_offset(),
             trailing_offset: order.trailing_offset(),